refactor: add router component
