+++ firmware/components/um1_lan/CMakeLists.txt
@@ -1,5 +1,5 @@
 idf_component_register(SRCS um1_lan.c start_lan.c ethernet_init.c
                     INCLUDE_DIRS "include"
-                    REQUIRES esp_netif esp_eth esp_event um1_spiffs um1_config um1_uart
+                    REQUIRES esp_netif esp_eth esp_event um1_spiffs um1_config um1_router driver
                     )
