+++ firmware/components/um1_router/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "um1_router.c"
+                    INCLUDE_DIRS "include"
+                    REQUIRES driver um1_config um1_mqtt)
