+++ firmware/components/um1_router/include/um1_router.h
@@ -1,0 +1,9 @@
+#ifndef UM1_ROUTER_H
+#define UM1_ROUTER_H
+
+#include <stddef.h>
+#include <stdint.h>
+
+void route_data(const char *src_if, const uint8_t *data, size_t len);
+
+#endif // UM1_ROUTER_H
