+++ firmware/components/um1_uart/CMakeLists.txt
@@ -1,5 +1,5 @@
 idf_component_register(SRCS "um1_uart.c"
                     INCLUDE_DIRS "include"
-                    REQUIRES driver um1_http_server um1_mqtt um1_sntp um1_config
+                    REQUIRES driver um1_http_server um1_sntp um1_config um1_router
                     )
