+++ firmware/components/um1_wifi/CMakeLists.txt
@@ -1,5 +1,5 @@
 idf_component_register(SRCS um1_wifi.c
                     INCLUDE_DIRS "include"
-                    REQUIRES esp_wifi nvs_flash um1_config um1_uart
+                    REQUIRES esp_wifi nvs_flash um1_config um1_router
                     )
