+++ firmware/components/um1_wifi/um1_wifi.c
@@ -1,5 +1,5 @@
 #include "um1_wifi.h"
-#include "um1_uart.h"
+#include "um1_router.h"
 
 #define TAG "um1_wifi"
 #define TCP_PORT 4444
